+++ src/Infrastructure/ui/slack/lib/ResponseParser.ts
